+++ servers/Azure.Mcp.Server/CHANGELOG.md
@@ -15,6 +15,8 @@ The Azure MCP Server updates automatically by default whenever a new release com
 
 - Fixed problem where help option (`--help`) was showing an error and enabling it across all commands and command groups. [[#583](https://github.com/microsoft/mcp/pull/583)]
 
+- Fixed `azmcp_kusto_table_schema` to return table schema. [[#530](https://github.com/microsoft/mcp/issues/530)]
+
 ### Other Changes
 
 ## 0.8.2 (2025-09-25)